Use DeleteChatResourcesAsync in DeleteChatSessionHistoryAsync endpoint

+++ webapi/Controllers/ChatHistoryController.cs
@@ -381,7 +381,7 @@ public class ChatHistoryController : ControllerBase
         // Delete any resources associated with the chat session.
         try
         {
-            await this.DeleteChatResourcesAsync(chatIdString, cancellationToken);
+            await this.DeleteChatResourcesAsync(chatIdString, true, cancellationToken);
         }
         catch (AggregateException)
         {
@@ -433,11 +433,7 @@ public class ChatHistoryController : ControllerBase
         // Delete chat messages in chat session.
         try
         {
-            var messages = await this._messageRepository.FindByChatIdAsync(chatIdString);
-            foreach (var message in messages)
-            {
-                await this._messageRepository.DeleteAsync(message);
-            }
+            await this.DeleteChatResourcesAsync(chatIdString, false, cancellationToken);
 
             // Create deleted chat history bot message
             var chatMessage = CopilotChatMessage.CreateBotResponseMessage(
@@ -470,15 +466,19 @@ public class ChatHistoryController : ControllerBase
     /// Deletes all associated resources (messages, memories, participants) associated with a chat session.
     /// </summary>
     /// <param name="chatId">The chat id.</param>
-    private async Task DeleteChatResourcesAsync(string chatId, CancellationToken cancellationToken)
+    /// <param name="shouldDeleteParticipants">Flag to determine if participants should be deleted.</param>
+    private async Task DeleteChatResourcesAsync(string chatId, bool shouldDeleteParticipants, CancellationToken cancellationToken)
     {
         var cleanupTasks = new List<Task>();
 
-        // Create and store the tasks for deleting all users tied to the chat.
-        var participants = await this._participantRepository.FindByChatIdAsync(chatId);
-        foreach (var participant in participants)
+        if (shouldDeleteParticipants)
         {
-            cleanupTasks.Add(this._participantRepository.DeleteAsync(participant));
+            // Create and store the tasks for deleting all users tied to the chat.
+            var participants = await this._participantRepository.FindByChatIdAsync(chatId);
+            foreach (var participant in participants)
+            {
+                cleanupTasks.Add(this._participantRepository.DeleteAsync(participant));
+            }
         }
 
         // Create and store the tasks for deleting chat messages.
